ocr/requirements_easyocr.txt
# Flask
Flask==3.0.0
flask-cors==4.0.0

# EasyOCR (lightweight alternative to PaddleOCR)
easyocr==1.7.0

# Image processing
Pillow==10.2.0
numpy==1.24.3

# Optional: For PDF support
PyMuPDF==1.23.8

# Production server
gunicorn==21.2.0
